Completed Task #1, 2, and 3
Finally completed something usefull and remembered to upload it to github instead of leaving it on my machine. This code took me probbaly 5 hours to write.

I also delete the old main and string.h because that was c, and I will be using c++ as we are learning it in COMP III this semester.