Completed Task # 14-23
Worked a lot on it today, completed all tasks from 14 to 23, mainly just set building stuff, and a bunch of tests. all of my tests for the assignments are in order by DFA_TEST#. the # just represents what # they appeared in the homework. Im proud of half of my code, and the other half is disgusting but works. really tough to understand half of this with my limited c++ knowledge