Completed Task #23
Finished the constructor for the NFA class. made functions to get the data out of a DFA so i could convert a DFA into an NFA without having the classes yell at each other. spent a few hours with those errors, so this was the only solution that i could code.